pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mr_imagen"
version = "1.0.0"
authors = [
  { name = "MindRoot" },
]
description = "Image generation plugin using Google's Imagen API"
requires-python = ">=3.8"
dependencies = [
    "google-generativeai",
    "Pillow",
]
